float to long

+++ src/main/java/globesort/GlobeSortClient.java
@@ -54,7 +54,7 @@ public class GlobeSortClient {
         long t4 = System.currentTimeMillis();
         long applicationThroughput = (values.length*1000/(t4-t3));
         float networkThroughputMsecs = (t4-t3-response.getValues(values.length))/(2);
-        float networkThroughput = values.length*1000/networkThroughputMsecs;
+        long networkThroughput = values.length*1000/(long)networkThroughputMsecs;
         System.out.println("Sorted array received. Num Records: " +values.length+ ". Application Throughput(records/sec): " +applicationThroughput);
         System.out.println("Time taken by server to sort(msec): "
                 +response.getValues(values.length)+
